fix: 최초 접속시 copyCode 로드되지 않는 버그 수정
